requirements.txt
aiohttp==3.9.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.12.1
attrs==25.4.0
bcrypt==4.0.1
cachetools==5.3.2
certifi==2026.2.25
cffi==2.0.0
chardet==7.1.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
Deprecated==1.3.1
deprecation==2.1.0
distro==1.9.0
ecdsa==0.19.1
et_xmlfile==2.0.0
fastapi==0.109.0
filelock==3.25.2
frozenlist==1.8.0
fsspec==2026.2.0
gotrue==1.3.0
h11==0.14.0
h2==4.3.0
hf-xet==1.4.2
hpack==4.1.0
httpcore==0.17.3
httptools==0.7.1
httpx==0.27.0
huggingface_hub==1.7.1
hyperframe==6.1.0
idna==3.11
limits==5.8.0
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.1
numpy==1.26.4
openpyxl==3.1.2
packaging==26.0
pandas==2.2.2
passlib==1.7.4
pillow==12.1.1
postgrest==0.13.2
propcache==0.4.1
pyasn1==0.6.3
pycparser==3.0
pydantic==2.5.3
pydantic[email]==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.3
realtime==1.0.6
reportlab==4.1.0
requests==2.32.5
rich==14.3.3
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.35.1
storage3==0.6.1
StrEnum==0.4.15
stripe==8.0.0
supabase==2.7.4
supafunc==0.3.3
tenacity==8.2.3
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.27.0
watchfiles==1.1.1
websockets==12.0
wrapt==2.1.2
yarl==1.23.0
keepa==1.3.8
pytrends==4.9.2
scipy==1.11.4
google-generativeai==0.8.3
openai==1.30.5
redis==5.0.1
bleach==6.1.0
beautifulsoup4==4.12.3
lxml==5.1.0
spacy==3.7.4
catboost==1.2.5
apscheduler==3.10.4
plotly==5.22.0
scikit-learn==1.4.2